+++ Assets/src/MetaBlocks/MetaBlockType.cs
@@ -5,7 +5,7 @@ public class MetaBlockType : BlockType
     public readonly System.Type componentType;
     private readonly System.Type propertiesType;
 
-    public MetaBlockType(byte id, string name, System.Type componentType, System.Type propertiesType)
+    public MetaBlockType(int id, string name, System.Type componentType, System.Type propertiesType)
         : base(id, name, false, 0, 0, 0, 0, 0, 0)
     {
         this.componentType = componentType;
