+++ Assets/src/MetaBlocks/VideoBlock/VideoBlockType.cs
@@ -1,6 +1,6 @@
 public class VideoBlockType : MetaBlockType
 {
-    public VideoBlockType(byte id) : base(id, "video", typeof(VideoBlockObject), typeof(VideoBlockProperties))
+    public VideoBlockType(int id) : base(id, "video", typeof(VideoBlockObject), typeof(VideoBlockProperties))
     {
     }
 }
